# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login a registered user",
                "operationId": "login",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "The email of the user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "The password of the user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário ou senha inválido!"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "logout a registered user",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout foi realizado com sucesso!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product": {
            "get": {
                "tags": [
                    "product"
                ],
                "description": "Return all products",
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "product"
                ],
                "description": "Create Product.",
                "operationId": "1d0b0d8724685b7b8f9d87f403bfc4df",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "price",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/{id}": {
            "put": {
                "tags": [
                    "product"
                ],
                "description": "Edit product.",
                "operationId": "d3467c795490afbb49680821c9bd6b0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "price",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/{product}": {
            "get": {
                "tags": [
                    "product"
                ],
                "description": "Return product",
                "operationId": "5c1d0ea045a25999c87076261c7af02e",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "identificador do produto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "product"
                ],
                "description": "Delete product",
                "operationId": "e32000a87befe10b46301e373b603542",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "identificador do produto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "unable delete"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "description": "Return all users",
                "operationId": "971db9bc974c0e607288cfe7ecc10096",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{user}": {
            "get": {
                "tags": [
                    "user"
                ],
                "description": "Return user",
                "operationId": "a0fa3a08f20fb611866a52d1bb2e43dd",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "identificador do usuário",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "description": "Delete user",
                "operationId": "887f11e9971bf8f135d6acd66435514a",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "identificador do usuário",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "unable delete"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}